2.2 update summary table

+++ lib/model/division_summary.dart
@@ -1,12 +1,12 @@
 class DivisionSummary {
   final String division;
 
-  final int allTtl, allI, allII, allIII, allIV, allV;
+  final double allTtl, allI, allII, allIII, allIV, allV;
 
-  final int proDoneTtl, proDoneI, proDoneII, proDoneIII, proDoneIV, proDoneV;
-  final int hseDoneTtl, hseDoneI, hseDoneII, hseDoneIII, hseDoneIV, hseDoneV;
+  final double proDoneTtl, proDoneI, proDoneII, proDoneIII, proDoneIV, proDoneV;
+  final double hseDoneTtl, hseDoneI, hseDoneII, hseDoneIII, hseDoneIV, hseDoneV;
 
-  final int remainTtl, remainI, remainII, remainIII, remainIV, remainV;
+  final double remainTtl, remainI, remainII, remainIII, remainIV, remainV;
 
   DivisionSummary({
     required this.division,
@@ -38,29 +38,29 @@ class DivisionSummary {
 
   factory DivisionSummary.fromJson(Map<String, dynamic> j) => DivisionSummary(
     division: (j['division'] ?? '').toString(),
-    allTtl: (j['allTtl'] ?? 0) as int,
-    allI: (j['allI'] ?? 0) as int,
-    allII: (j['allII'] ?? 0) as int,
-    allIII: (j['allIII'] ?? 0) as int,
-    allIV: (j['allIV'] ?? 0) as int,
-    allV: (j['allV'] ?? 0) as int,
-    proDoneTtl: (j['proDoneTtl'] ?? 0) as int,
-    proDoneI: (j['proDoneI'] ?? 0) as int,
-    proDoneII: (j['proDoneII'] ?? 0) as int,
-    proDoneIII: (j['proDoneIII'] ?? 0) as int,
-    proDoneIV: (j['proDoneIV'] ?? 0) as int,
-    proDoneV: (j['proDoneV'] ?? 0) as int,
-    hseDoneTtl: (j['hseDoneTtl'] ?? 0) as int,
-    hseDoneI: (j['hseDoneI'] ?? 0) as int,
-    hseDoneII: (j['hseDoneII'] ?? 0) as int,
-    hseDoneIII: (j['hseDoneIII'] ?? 0) as int,
-    hseDoneIV: (j['hseDoneIV'] ?? 0) as int,
-    hseDoneV: (j['hseDoneV'] ?? 0) as int,
-    remainTtl: (j['remainTtl'] ?? 0) as int,
-    remainI: (j['remainI'] ?? 0) as int,
-    remainII: (j['remainII'] ?? 0) as int,
-    remainIII: (j['remainIII'] ?? 0) as int,
-    remainIV: (j['remainIV'] ?? 0) as int,
-    remainV: (j['remainV'] ?? 0) as int,
+    allTtl: (j['allTtl'] ?? 0) as double,
+    allI: (j['allI'] ?? 0) as double,
+    allII: (j['allII'] ?? 0) as double,
+    allIII: (j['allIII'] ?? 0) as double,
+    allIV: (j['allIV'] ?? 0) as double,
+    allV: (j['allV'] ?? 0) as double,
+    proDoneTtl: (j['proDoneTtl'] ?? 0) as double,
+    proDoneI: (j['proDoneI'] ?? 0) as double,
+    proDoneII: (j['proDoneII'] ?? 0) as double,
+    proDoneIII: (j['proDoneIII'] ?? 0) as double,
+    proDoneIV: (j['proDoneIV'] ?? 0) as double,
+    proDoneV: (j['proDoneV'] ?? 0) as double,
+    hseDoneTtl: (j['hseDoneTtl'] ?? 0) as double,
+    hseDoneI: (j['hseDoneI'] ?? 0) as double,
+    hseDoneII: (j['hseDoneII'] ?? 0) as double,
+    hseDoneIII: (j['hseDoneIII'] ?? 0) as double,
+    hseDoneIV: (j['hseDoneIV'] ?? 0) as double,
+    hseDoneV: (j['hseDoneV'] ?? 0) as double,
+    remainTtl: (j['remainTtl'] ?? 0) as double,
+    remainI: (j['remainI'] ?? 0) as double,
+    remainII: (j['remainII'] ?? 0) as double,
+    remainIII: (j['remainIII'] ?? 0) as double,
+    remainIV: (j['remainIV'] ?? 0) as double,
+    remainV: (j['remainV'] ?? 0) as double,
   );
 }
